Moved the steps to start and stop the callback into the same scenario as checking the case details.
