+++ behave/features/cla_select_digital_case_by_call_back.feature
@@ -16,5 +16,10 @@ Scenario: Can see details of selected callback
     | details     |
     | Full name   |
     | Telephone   |
+    When I select "Start Call"
+    Then the call has started
+    And I remove the callback
+    Then this case is removed from callback list (calendar view)
+
